+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers.Example.NUnit/TestTemplate.cs
@@ -1,8 +1,5 @@
-﻿using kCura.Relativity.Client;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using Relativity.API;
-using Relativity.Services.ServiceProxy;
-using Relativity.Test.Helpers.ServiceFactory.Extentions;
 using Relativity.Test.Helpers.SharedTestHelpers;
 using System;
 using System.IO;
@@ -16,8 +13,8 @@ namespace Relativity.Test.Helpers.Example.NUnit
 	/// 
 	/// Relativity Integration Test Helpers to assist you with writing good Integration Tests for your application. You can use this framework to test event handlers, agents and any workflow that combines agents and frameworks.
 	/// 
-	/// Before you get Started, fill out details for the following the app.config file
-	/// "WorkspaceID", "RSAPIServerAddress", "RESTServerAddress",	"AdminUsername","AdminPassword", "SQLServerAddress" ,"SQLUsername","SQLPassword" "TestWorkspaceName"
+	/// Before you get Started, fill out details for the following the app.config file or runsettings file
+	/// "WorkspaceID", "RelativityInstanceAddress", "RESTServerAddress",	"AdminUsername","AdminPassword", "SQLServerAddress" ,"SQLUsername","SQLPassword" "TestWorkspaceName"
 	/// 
 	/// </summary>
 
@@ -26,7 +23,6 @@ namespace Relativity.Test.Helpers.Example.NUnit
 	{
 
 		#region Variables
-		private IRSAPIClient _client;
 		private int _workspaceId;
 		private Int32 _rootFolderArtifactID;
 		private string _workspaceName = $"IntTest_{Guid.NewGuid()}";
@@ -41,7 +37,6 @@ namespace Relativity.Test.Helpers.Example.NUnit
 		private int _longtextartid;
 		private int _yesnoartid;
 		private int _wholeNumberArtId;
-		private Services.ServiceProxy.ServiceFactory _serviceFactory;
 
 		#endregion
 
@@ -55,23 +50,16 @@ namespace Relativity.Test.Helpers.Example.NUnit
 			var helper = new TestHelper(ConfigurationHelper.ADMIN_USERNAME, ConfigurationHelper.DEFAULT_PASSWORD);
 			_servicesMgr = helper.GetServicesManager();
 
-			// implement_IHelper
-			//create client
-			_client = helper.GetServicesManager().GetProxy<IRSAPIClient>(ConfigurationHelper.ADMIN_USERNAME, ConfigurationHelper.DEFAULT_PASSWORD);
-
-			//use helper method to instantiate a new service factory
-			_serviceFactory = GetServiceFactory();
-
 			//Create new user 
-			_userArtifactId = Relativity.Test.Helpers.UserHelpers.CreateUser.CreateNewUser(_client);
+			_userArtifactId = Relativity.Test.Helpers.UserHelpers.UserHelper.Create(helper.GetServicesManager());
 
 			//Create new group
-			Relativity.Test.Helpers.GroupHelpers.GroupHelper.CreateGroup(_serviceFactory, _groupName);
+			Relativity.Test.Helpers.GroupHelpers.GroupHelper.CreateGroup(_servicesMgr, _groupName);
 
 
 			//Create workspace
 			_workspaceId = Helpers.WorkspaceHelpers.WorkspaceHelpers.CreateAsync(_servicesMgr, _workspaceName, SharedTestHelpers.ConfigurationHelper.TEST_WORKSPACE_TEMPLATE_NAME).ConfigureAwait(false).GetAwaiter().GetResult();
-			_client.APIOptions.WorkspaceID = _workspaceId;
+
 			var executableLocation = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
 			var nativeFilePath = "";
 			var nativeName = @"\\\\FakeFilePath\Natives\SampleTextFile.txt";
@@ -87,18 +75,18 @@ namespace Relativity.Test.Helpers.Example.NUnit
 			Relativity.Test.Helpers.ImportAPIHelper.ImportAPIHelper.CreateDocumentswithFolderName(_workspaceId, _numberOfDocuments, folderName, nativeFilePath);
 
 			//Create Fixed Length field
-			_fixedLengthArtId = Relativity.Test.Helpers.ArtifactHelpers.FieldHelper.CreateFieldFixedLengthText(_serviceFactory, _workspaceId);
+			_fixedLengthArtId = Relativity.Test.Helpers.ArtifactHelpers.FieldHelper.CreateFieldFixedLengthText(_servicesMgr, _workspaceId);
 
 			//Create Long Text Field
-			_longtextartid = Relativity.Test.Helpers.ArtifactHelpers.FieldHelper.CreateFieldLongText(_serviceFactory, _workspaceId);
+			_longtextartid = Relativity.Test.Helpers.ArtifactHelpers.FieldHelper.CreateFieldLongText(_servicesMgr, _workspaceId);
 
 			//Create Whole number field
-			_wholeNumberArtId = Relativity.Test.Helpers.ArtifactHelpers.FieldHelper.CreateFieldWholeNumber(_serviceFactory, _workspaceId);
+			_wholeNumberArtId = Relativity.Test.Helpers.ArtifactHelpers.FieldHelper.CreateFieldWholeNumber(_servicesMgr, _workspaceId);
 
 			_workspaceId = 1017834;
 			var DtSearchAppArtifactId = 1038135;
 
-			var guid = helper.GetGuid(_workspaceId, DtSearchAppArtifactId);
+			var guid = helper.GetGuid(_workspaceId, DtSearchAppArtifactId, Constants.ArtifactTypeIds.Search);
 		}
 
 		#endregion
@@ -113,12 +101,10 @@ namespace Relativity.Test.Helpers.Example.NUnit
 			WorkspaceHelpers.WorkspaceHelpers.Delete(_servicesMgr, _workspaceId);
 
 			//Delete User
-			UserHelpers.DeleteUser.Delete_User(_client, _userArtifactId);
+			UserHelpers.UserHelper.Delete(_servicesMgr, _userArtifactId);
 
 			//Delete Group
-			GroupHelpers.GroupHelper.DeleteGroup(_serviceFactory, _groupArtifactId);
-
-			_serviceFactory = null;
+			GroupHelpers.GroupHelper.DeleteGroup(_servicesMgr, _groupArtifactId);
 		}
 
 
@@ -139,26 +125,6 @@ namespace Relativity.Test.Helpers.Example.NUnit
 
 		#endregion
 
-		//helper method to create a service factory
-		private Services.ServiceProxy.ServiceFactory GetServiceFactory()
-		{
-			var relativityServicesUri = new Uri($"{ConfigurationHelper.SERVER_BINDING_TYPE}://{ConfigurationHelper.RELATIVITY_INSTANCE_ADDRESS}/Relativity.Services");
-			var relativityRestUri = new Uri($"{ConfigurationHelper.SERVER_BINDING_TYPE}://{ConfigurationHelper.REST_SERVER_ADDRESS.ToLower().Replace("-services", "")}/Relativity.Rest/Api");
-
-			Relativity.Services.ServiceProxy.UsernamePasswordCredentials usernamePasswordCredentials = new Relativity.Services.ServiceProxy.UsernamePasswordCredentials(
-				username: ConfigurationHelper.ADMIN_USERNAME,
-				password: ConfigurationHelper.DEFAULT_PASSWORD);
-
-			ServiceFactorySettings serviceFactorySettings = new ServiceFactorySettings(
-				relativityServicesUri: relativityServicesUri,
-				relativityRestUri: relativityRestUri,
-				credentials: usernamePasswordCredentials);
-
-			var serviceFactory = new Services.ServiceProxy.ServiceFactory(
-				settings: serviceFactorySettings);
-
-			return serviceFactory;
-		}
 	}
 
 }