+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/ArtifactHelpers/ClientHelper.cs
@@ -1,18 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using Relativity.API;
+using Relativity.Services;
 using Relativity.Services.Choice;
 using Relativity.Services.Client;
-using Relativity.Test.Helpers.ArtifactHelpers.Interfaces;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace Relativity.Test.Helpers.ArtifactHelpers
 {
 	public class ClientHelper
 	{
-		public static int CreateClient(Services.ServiceProxy.ServiceFactory serviceFactory, string name)
+		public static int CreateClient(IServicesMgr servicesMgr, string name)
 		{
 			try
 			{
-				using (IClientManager proxy = serviceFactory.CreateProxy<IClientManager>())
+				using (IClientManager proxy = servicesMgr.CreateProxy<IClientManager>(ExecutionIdentity.CurrentUser))
 				{
 					List<ChoiceRef> choiceRefs = proxy.GetStatusChoicesForClientAsync().ConfigureAwait(false).GetAwaiter().GetResult();
 					ChoiceRef statusRef = choiceRefs.Find(x => x.Name == "Active");
@@ -21,7 +24,8 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 					{
 						Name = name,
 						Number = Guid.NewGuid().ToString(),
-						Status = statusRef, Keywords = "Test Client",
+						Status = statusRef,
+						Keywords = "Test Client",
 						Notes = "Created with Relativity Test Helpers."
 					};
 
@@ -36,11 +40,11 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 			}
 		}
 
-		public static void DeleteClient(Services.ServiceProxy.ServiceFactory serviceFactory, int artifactId)
+		public static void DeleteClient(IServicesMgr servicesMgr, int artifactId)
 		{
 			try
 			{
-				using (IClientManager proxy = serviceFactory.CreateProxy<IClientManager>())
+				using (IClientManager proxy = servicesMgr.CreateProxy<IClientManager>(ExecutionIdentity.CurrentUser))
 				{
 					proxy.DeleteSingleAsync(artifactId).ConfigureAwait(false).GetAwaiter().GetResult();
 				}
@@ -49,6 +53,23 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 			{
 				throw new Exception("Unable to delete client.", ex);
 			}
+		}
+
+		public static async Task<int> GetClientId(IServicesMgr servicesMgr, string clientName)
+		{
+			int clientId;
+
+			using (IClientManager clientManager = servicesMgr.CreateProxy<IClientManager>(ExecutionIdentity.CurrentUser))
+			{
+				Services.Query query = new Services.Query()
+				{
+					Condition = new TextCondition("Name", TextConditionEnum.EqualTo, clientName).ToQueryString()
+				};
+				ClientQueryResultSet result = await clientManager.QueryAsync(query);
+				clientId = result.Results.First().Artifact.ArtifactID;
+			}
+
+			return clientId;
 		}
 	}
 }