+++ Source/Relativity.Test.Helpers/Relativity.Test.Helpers/ArtifactHelpers/FieldHelper.cs
@@ -1,21 +1,16 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
-using System.Data.SqlClient;
-using System.Linq;
-using kCura.Vendor.Castle.Core.Internal;
-using Newtonsoft.Json;
-using Relativity.Test.Helpers.ArtifactHelpers.Interfaces;
-using Relativity.Test.Helpers.Exceptions;
-using TestHelpersKepler.Interfaces.TestHelpersModule.v1.Models;
+﻿using kCura.Vendor.Castle.Core.Internal;
+using Relativity.API;
 using Relativity.Services.Interfaces.Field;
 using Relativity.Services.Interfaces.Field.Models;
 using Relativity.Services.Interfaces.Shared.Models;
-using Relativity.API;
 using Relativity.Services.Objects;
 using Relativity.Services.Objects.DataContracts;
+using Relativity.Test.Helpers.ArtifactHelpers.Interfaces;
 using Relativity.Test.Helpers.ServiceFactory.Extentions;
 using Relativity.Test.Helpers.SharedTestHelpers;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using FieldType = Relativity.Services.Interfaces.Field.Models.FieldType;
 
 namespace Relativity.Test.Helpers.ArtifactHelpers
@@ -95,7 +90,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 			}
 		}
 
-		public static int CreateFieldDate(Services.ServiceProxy.ServiceFactory serviceFactory, int workspaceID)
+		public static int CreateFieldDate(IServicesMgr servicesMgr, int workspaceID)
 		{
 			try
 			{
@@ -120,7 +115,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 					Notes = "Created by Test Helpers"
 				};
 
-				using (IFieldManager fieldManager = serviceFactory.CreateProxy<IFieldManager>())
+				using (IFieldManager fieldManager = servicesMgr.CreateProxy<IFieldManager>(ExecutionIdentity.CurrentUser))
 				{
 					fieldId = fieldManager.CreateDateFieldAsync(workspaceID, fieldRequest).ConfigureAwait(false).GetAwaiter().GetResult();
 				}
@@ -133,7 +128,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 			}
 		}
 
-		public static int CreateFieldUser(Services.ServiceProxy.ServiceFactory serviceFactory, int workspaceID)
+		public static int CreateFieldUser(IServicesMgr servicesMgr, int workspaceID)
 		{
 			try
 			{
@@ -156,7 +151,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 					Notes = "Created by Test Helpers"
 				};
 
-				using (IFieldManager fieldManager = serviceFactory.CreateProxy<IFieldManager>())
+				using (IFieldManager fieldManager = servicesMgr.CreateProxy<IFieldManager>(ExecutionIdentity.CurrentUser))
 				{
 					fieldId = fieldManager.CreateUserFieldAsync(workspaceID, fieldRequest).ConfigureAwait(false).GetAwaiter().GetResult();
 				}
@@ -169,7 +164,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 			}
 		}
 
-		public static int CreateFieldFixedLengthText(Services.ServiceProxy.ServiceFactory serviceFactory, int workspaceID)
+		public static int CreateFieldFixedLengthText(IServicesMgr servicesMgr, int workspaceID)
 		{
 			try
 			{
@@ -197,7 +192,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 					Notes = "Created by Test Helpers"
 				};
 
-				using (IFieldManager fieldManager = serviceFactory.CreateProxy<IFieldManager>())
+				using (IFieldManager fieldManager = servicesMgr.CreateProxy<IFieldManager>(ExecutionIdentity.CurrentUser))
 				{
 					fieldId = fieldManager.CreateFixedLengthFieldAsync(workspaceID, fieldRequest).ConfigureAwait(false).GetAwaiter().GetResult();
 				}
@@ -209,7 +204,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 				throw new Exception("Error creating field.", ex);
 			}
 		}
-		public static int CreateFieldLongText(Services.ServiceProxy.ServiceFactory serviceFactory, int workspaceID)
+		public static int CreateFieldLongText(IServicesMgr servicesMgr, int workspaceID)
 		{
 			try
 			{
@@ -236,7 +231,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 					Notes = "Created by Test Helpers"
 				};
 
-				using (IFieldManager fieldManager = serviceFactory.CreateProxy<IFieldManager>())
+				using (IFieldManager fieldManager = servicesMgr.CreateProxy<IFieldManager>(ExecutionIdentity.CurrentUser))
 				{
 					fieldId = fieldManager.CreateLongTextFieldAsync(workspaceID, fieldRequest).ConfigureAwait(false).GetAwaiter().GetResult();
 				}
@@ -248,7 +243,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 				throw new Exception("Error creating field.", ex);
 			}
 		}
-		public static int CreateFieldWholeNumber(Services.ServiceProxy.ServiceFactory serviceFactory, int workspaceID)
+		public static int CreateFieldWholeNumber(IServicesMgr servicesMgr, int workspaceID)
 		{
 			try
 			{
@@ -271,7 +266,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 					Notes = "Created by Test Helpers"
 				};
 
-				using (IFieldManager fieldManager = serviceFactory.CreateProxy<IFieldManager>())
+				using (IFieldManager fieldManager = servicesMgr.CreateProxy<IFieldManager>(ExecutionIdentity.CurrentUser))
 				{
 					fieldId = fieldManager.CreateWholeNumberFieldAsync(workspaceID, fieldRequest).ConfigureAwait(false).GetAwaiter().GetResult();
 				}
@@ -283,7 +278,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 				throw new Exception("Error creating field.", ex);
 			}
 		}
-		public static int CreateFieldYesNo(Services.ServiceProxy.ServiceFactory serviceFactory, int workspaceID)
+		public static int CreateFieldYesNo(IServicesMgr servicesMgr, int workspaceID)
 		{
 			try
 			{
@@ -308,7 +303,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 					Notes = "Created by Test Helpers"
 				};
 
-				using (IFieldManager fieldManager = serviceFactory.CreateProxy<IFieldManager>())
+				using (IFieldManager fieldManager = servicesMgr.CreateProxy<IFieldManager>(ExecutionIdentity.CurrentUser))
 				{
 					fieldId = fieldManager.CreateYesNoFieldAsync(workspaceID, fieldRequest).ConfigureAwait(false).GetAwaiter().GetResult();
 				}
@@ -320,7 +315,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 				throw new Exception("Error creating field.", ex);
 			}
 		}
-		public static int CreateFieldSingleChoice(Services.ServiceProxy.ServiceFactory serviceFactory, int workspaceID)
+		public static int CreateFieldSingleChoice(IServicesMgr servicesMgr, int workspaceID)
 		{
 			try
 			{
@@ -344,7 +339,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 					Notes = "Created by Test Helpers"
 				};
 
-				using (IFieldManager fieldManager = serviceFactory.CreateProxy<IFieldManager>())
+				using (IFieldManager fieldManager = servicesMgr.CreateProxy<IFieldManager>(ExecutionIdentity.CurrentUser))
 				{
 					fieldId = fieldManager.CreateSingleChoiceFieldAsync(workspaceID, fieldRequest).ConfigureAwait(false).GetAwaiter().GetResult();
 				}
@@ -356,7 +351,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 				throw new Exception("Error creating field.", ex);
 			}
 		}
-		public static int CreateFieldMultipleChoice(Services.ServiceProxy.ServiceFactory serviceFactory, int workspaceID)
+		public static int CreateFieldMultipleChoice(IServicesMgr servicesMgr, int workspaceID)
 		{
 			try
 			{
@@ -380,7 +375,7 @@ namespace Relativity.Test.Helpers.ArtifactHelpers
 					Notes = "Created by Test Helpers"
 				};
 
-				using (IFieldManager fieldManager = serviceFactory.CreateProxy<IFieldManager>())
+				using (IFieldManager fieldManager = servicesMgr.CreateProxy<IFieldManager>(ExecutionIdentity.CurrentUser))
 				{
 					fieldId = fieldManager.CreateMultipleChoiceFieldAsync(workspaceID, fieldRequest).ConfigureAwait(false).GetAwaiter().GetResult();
 				}